requirements.txt
streamlit>=1.19.0
streamlit-audiorecorder
# audio-recorder-streamlit
git+https://github.com/openai/whisper.git
